+++ src/lwpa/netint.c
@@ -86,7 +86,21 @@ size_t lwpa_netint_get_num_interfaces()
   return (init_count ? netint_cache.num_netints : 0);
 }
 
-/*! \brief Enumerate the network interfaces on the system.
+/*! \brief Get a list of network interfaces on the system.
+ *
+ *  For NICs with multiple IP addresses assigned, this module separates each address into its own
+ *  entry in the netint array. Because of this, multiple array entries could have the same value
+ *  for the index, mac and name parameters.
+ *
+ *  \return Pointer to an array of network interfaces of length lwpa_netint_get_num_interfaces(),
+ *          or NULL if there are no interfaces present or the module is not initialized.
+ */
+const LwpaNetintInfo* lwpa_netint_get_interfaces()
+{
+  return (init_count ? netint_cache.netints : NULL);
+}
+
+/*! \brief Copy the list of network interfaces on the system into an array.
  *
  *  For NICs with multiple IP addresses assigned, this module separates each address into its own
  *  entry in the netint array. Because of this, multiple array entries could have the same value
@@ -95,10 +109,10 @@ size_t lwpa_netint_get_num_interfaces()
  *  \param[out] netint_arr Array of network interface description structs to fill in with interface
  *                         info.
  *  \param[in] netint_arr_size Size of the netint array.
- *  \return Number of network interfaces that were enumerated, up to a maximum of netint_arr_size,
- *          or 0 if there are no interfaces present or an error occurred.
+ *  \return Number of network interfaces that were copied, up to a maximum of netint_arr_size,
+ *          or 0 if there are no interfaces present or the module is not initialized.
  */
-size_t lwpa_netint_get_interfaces(LwpaNetintInfo* netint_arr, size_t netint_arr_size)
+size_t lwpa_netint_copy_interfaces(LwpaNetintInfo* netint_arr, size_t netint_arr_size)
 {
   if (!init_count || !netint_arr || netint_arr_size == 0)
     return 0;