+++ tests/unit/lwpa/test_netint.cpp
@@ -39,19 +39,30 @@ protected:
 TEST_F(NetintTest, enumerate)
 {
   ASSERT_GT(num_netints, 0u);
-  auto netint_arr = std::make_unique<LwpaNetintInfo[]>(num_netints);
-  size_t num_netints_returned = lwpa_netint_get_interfaces(netint_arr.get(), num_netints);
-  ASSERT_EQ(num_netints_returned, num_netints);
 
   size_t num_defaults = 0;
-  for (LwpaNetintInfo* netint = netint_arr.get(); netint < netint_arr.get() + num_netints_returned; ++netint)
+  const LwpaNetintInfo* netint_list = lwpa_netint_get_interfaces();
+  ASSERT_NE(netint_list, nullptr);
+  for (auto netint = netint_list; netint < netint_list + num_netints; ++netint)
   {
     if (netint->is_default)
       ++num_defaults;
-    ASSERT_GT(strlen(netint->name), 0u);
+    EXPECT_GT(strlen(netint->name), 0u);
+    EXPECT_GT(strlen(netint->friendly_name), 0u);
   }
   // There can be a maximum of two default interfaces: one each for IPv4 and IPv6.
-  ASSERT_TRUE(num_defaults <= 2);
+  EXPECT_LE(num_defaults, 2u);
+}
+
+TEST_F(NetintTest, copy)
+{
+  // Test copying the full array
+  auto netint_arr = std::make_unique<LwpaNetintInfo[]>(num_netints);
+  size_t num_netints_returned = lwpa_netint_copy_interfaces(netint_arr.get(), num_netints);
+  ASSERT_EQ(num_netints_returned, num_netints);
+
+  // Make sure the array is equal to the list obtained from lwpa_netint_get_interfaces()
+  EXPECT_EQ(0, memcmp(netint_arr.get(), lwpa_netint_get_interfaces(), num_netints));
 }
 
 TEST_F(NetintTest, default)
@@ -62,11 +73,6 @@ TEST_F(NetintTest, default)
   LwpaNetintInfo def_v6;
   memset(&def_v4, 0, sizeof def_v4);
   memset(&def_v6, 0, sizeof def_v6);
-
-  auto netint_arr = std::make_unique<LwpaNetintInfo[]>(num_netints);
-  memset(netint_arr.get(), 0, sizeof(struct LwpaNetintInfo) * num_netints);
-
-  num_netints = lwpa_netint_get_interfaces(netint_arr.get(), num_netints);
 
   bool have_default_v4 = (kLwpaErrOk == lwpa_netint_get_default_interface(kLwpaIpTypeV4, &def_v4));
   bool have_default_v6 = (kLwpaErrOk == lwpa_netint_get_default_interface(kLwpaIpTypeV6, &def_v6));
@@ -80,7 +86,9 @@ TEST_F(NetintTest, default)
     EXPECT_TRUE(def_v6.is_default);
   }
 
-  for (LwpaNetintInfo* netint = netint_arr.get(); netint < netint_arr.get() + num_netints; ++netint)
+  const LwpaNetintInfo* netint_list = lwpa_netint_get_interfaces();
+  ASSERT_NE(netint_list, nullptr);
+  for (auto netint = netint_list; netint < netint_list + num_netints; ++netint)
   {
     if (netint->is_default)
     {
@@ -100,13 +108,13 @@ TEST_F(NetintTest, ipv4_routing)
 {
   ASSERT_GT(num_netints, 0u);
 
-  auto netint_arr = std::make_unique<LwpaNetintInfo[]>(num_netints);
-  num_netints = lwpa_netint_get_interfaces(netint_arr.get(), num_netints);
+  const LwpaNetintInfo* netint_list = lwpa_netint_get_interfaces();
+  ASSERT_NE(netint_list, nullptr);
 
   // For each normally routable (non-loopback, non-link-local) network interface, check to make sure
   // that lwpa_netint_get_interface_for_dest() resolves to that interface when asked for a route to
   // the interface address itself.
-  for (LwpaNetintInfo* netint = netint_arr.get(); netint < netint_arr.get() + num_netints; ++netint)
+  for (auto netint = netint_list; netint < netint_list + num_netints; ++netint)
   {
     if (!LWPA_IP_IS_V4(&netint->addr) || lwpa_ip_is_loopback(&netint->addr) || lwpa_ip_is_link_local(&netint->addr))
       continue;
@@ -128,6 +136,6 @@ TEST_F(NetintTest, ipv4_routing)
   LwpaIpAddr ext_addr;
   LWPA_IP_SET_V4_ADDRESS(&ext_addr, 0xc8dc0302);  // 200.220.3.2
   LwpaNetintInfo def;
-  ASSERT_EQ(kLwpaErrOk, lwpa_netint_get_interface_for_dest(&ext_addr, &def));
-  ASSERT_TRUE(def.is_default);
+  EXPECT_EQ(kLwpaErrOk, lwpa_netint_get_interface_for_dest(&ext_addr, &def));
+  EXPECT_TRUE(def.is_default);
 }